refactor(turns): split endturncontrol into focused turn ui modules
Split EndTurnControl into focused turn feature modules, such as control container, readiness
metrics, confirmation dialog, success message, and error/description helpers.

fix #122
