+++ src/features/turns/components/EndTurnControl.tsx
@@ -1,5 +1,5 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { CircleCheck, StepForward, TriangleAlert, X } from "lucide-react";
+import { StepForward } from "lucide-react";
 import { useState } from "react";
 
 import { ErrorState } from "@/components/shared/ErrorState";
@@ -8,13 +8,20 @@ import { Button } from "@/components/ui/button";
 import {
   formatSettlementReadinessPercentage,
   settlementReadinessSummaryQueryOptions,
-  type SettlementReadinessSummary,
 } from "@/features/settlements";
 
+import { endTurnBasicMutationOptions } from "../mutations/endTurnBasicMutations";
 import {
-  endTurnBasicMutationOptions,
-  isEndTurnBasicError,
-} from "../mutations/endTurnBasicMutations";
+  getControlDescription,
+  getErrorDescription,
+} from "../utils/endTurnDescriptions";
+
+import { EndTurnConfirmationDialog } from "./EndTurnConfirmationDialog";
+import { EndTurnMetric } from "./EndTurnMetric";
+import {
+  EndTurnSuccessMessage,
+  type SuccessfulEndTurnTransition,
+} from "./EndTurnSuccessMessage";
 
 import type { JSX } from "react";
 
@@ -26,12 +33,6 @@ type EndTurnControlProps = {
   readonly nextDateLabel: string;
   readonly nextTurnNumber: number;
   readonly worldId: string;
-};
-type SuccessfulEndTurnTransition = {
-  readonly nextDateLabel: string;
-  readonly nextTurnNumber: number;
-  readonly previousDateLabel: string;
-  readonly previousTurnNumber: number;
 };
 
 export function EndTurnControl({
@@ -219,223 +220,4 @@ function EndTurnControlContent({
       ) : null}
     </section>
   );
-}
-
-function EndTurnSuccessMessage({
-  transition,
-}: {
-  readonly transition: SuccessfulEndTurnTransition;
-}): JSX.Element {
-  return (
-    <div
-      className="grid gap-3 rounded-md border border-emerald-600/30 bg-emerald-50 px-3 py-3 text-sm text-emerald-950 dark:bg-emerald-950/30 dark:text-emerald-100"
-      role="status"
-    >
-      <p className="flex items-start gap-2 font-medium">
-        <CircleCheck className="mt-0.5 size-4 shrink-0" aria-hidden="true" />
-        End-turn transition completed.
-      </p>
-      <dl className="grid gap-2 sm:grid-cols-2">
-        <EndTurnDetail
-          label="Previous turn"
-          value={`Turn ${transition.previousTurnNumber}`}
-        />
-        <EndTurnDetail
-          label="New turn"
-          value={`Turn ${transition.nextTurnNumber}`}
-        />
-        <EndTurnDetail
-          label="Previous date"
-          value={transition.previousDateLabel}
-        />
-        <EndTurnDetail label="New date" value={transition.nextDateLabel} />
-      </dl>
-    </div>
-  );
-}
-
-function EndTurnDetail({
-  label,
-  value,
-}: {
-  readonly label: string;
-  readonly value: string;
-}): JSX.Element {
-  return (
-    <div>
-      <dt className="font-medium">{label}</dt>
-      <dd>{value}</dd>
-    </div>
-  );
-}
-
-function EndTurnConfirmationDialog({
-  currentDateLabel,
-  currentTurnNumber,
-  isPending,
-  nextDateLabel,
-  nextTurnNumber,
-  onCancel,
-  onConfirm,
-  readinessSummary,
-}: {
-  readonly currentDateLabel: string;
-  readonly currentTurnNumber: number;
-  readonly isPending: boolean;
-  readonly nextDateLabel: string;
-  readonly nextTurnNumber: number;
-  readonly onCancel: () => void;
-  readonly onConfirm: () => void;
-  readonly readinessSummary: SettlementReadinessSummary;
-}): JSX.Element {
-  const hasNotReadySettlements = readinessSummary.notReadySettlementCount > 0;
-
-  return (
-    <div className="fixed inset-0 z-50 grid place-items-center bg-background/80 p-4">
-      <div
-        aria-labelledby="end-turn-confirmation-title"
-        aria-modal="true"
-        className="grid w-full max-w-lg gap-5 rounded-md border border-border bg-card p-5 text-card-foreground shadow-lg"
-        role="dialog"
-      >
-        <div className="flex items-start justify-between gap-3">
-          <div className="space-y-1">
-            <h3
-              id="end-turn-confirmation-title"
-              className="text-lg font-semibold tracking-normal"
-            >
-              Confirm end turn
-            </h3>
-            <p className="text-sm text-muted-foreground">
-              This transition advances world state and cannot be undone.
-            </p>
-          </div>
-          <Button
-            aria-label="Cancel end turn"
-            disabled={isPending}
-            onClick={onCancel}
-            size="icon-sm"
-            type="button"
-            variant="ghost"
-          >
-            <X aria-hidden="true" />
-          </Button>
-        </div>
-
-        <dl className="grid gap-3 sm:grid-cols-2">
-          <EndTurnMetric
-            label="Current turn"
-            value={currentTurnNumber.toString()}
-          />
-          <EndTurnMetric label="Next turn" value={nextTurnNumber.toString()} />
-          <EndTurnMetric label="Current date" value={currentDateLabel} />
-          <EndTurnMetric label="Next date" value={nextDateLabel} />
-        </dl>
-
-        <div className="rounded-md border border-border bg-background px-3 py-2">
-          <p className="text-sm font-medium">Readiness summary</p>
-          <p className="text-sm text-muted-foreground">
-            {getReadinessSummaryDescription(readinessSummary)}
-          </p>
-        </div>
-
-        {hasNotReadySettlements ? (
-          <p
-            className="flex items-start gap-2 rounded-md border border-destructive/30 bg-destructive/10 px-3 py-2 text-sm text-destructive"
-            role="alert"
-          >
-            <TriangleAlert
-              className="mt-0.5 size-4 shrink-0"
-              aria-hidden="true"
-            />
-            Some settlements are not ready. You can still confirm and advance
-            the turn.
-          </p>
-        ) : null}
-
-        <div className="flex flex-col-reverse gap-2 sm:flex-row sm:justify-end">
-          <Button
-            disabled={isPending}
-            onClick={onCancel}
-            type="button"
-            variant="outline"
-          >
-            Cancel
-          </Button>
-          <Button disabled={isPending} onClick={onConfirm} type="button">
-            <StepForward aria-hidden="true" />
-            {isPending ? "Ending turn..." : "Confirm end turn"}
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function EndTurnMetric({
-  label,
-  value,
-}: {
-  readonly label: string;
-  readonly value: string;
-}): JSX.Element {
-  return (
-    <div className="rounded-md border border-border bg-background px-3 py-2">
-      <dt className="text-sm text-muted-foreground">{label}</dt>
-      <dd className="text-2xl font-semibold tracking-normal">{value}</dd>
-    </div>
-  );
-}
-
-function getControlDescription({
-  isArchived,
-  isPending,
-  isReadinessUnavailable,
-}: {
-  readonly isArchived: boolean;
-  readonly isPending: boolean;
-  readonly isReadinessUnavailable: boolean;
-}): string {
-  if (isArchived) {
-    return "End turn is disabled because this world is archived.";
-  }
-
-  if (isReadinessUnavailable) {
-    return "End turn is disabled until readiness can be reviewed.";
-  }
-
-  if (isPending) {
-    return "End-turn transition is running.";
-  }
-
-  return "Submitting starts one end-turn transition for the current turn.";
-}
-
-function getReadinessSummaryDescription(
-  readinessSummary: SettlementReadinessSummary,
-): string {
-  const readyPercentageLabel = formatSettlementReadinessPercentage(
-    readinessSummary.readyPercentage,
-  );
-
-  return `${readinessSummary.readySettlementCount} of ${readinessSummary.totalSettlementCount} settlements ready (${readyPercentageLabel}). ${readinessSummary.notReadySettlementCount} not ready.`;
-}
-
-function getErrorDescription(error: unknown): string {
-  if (isEndTurnBasicError(error)) {
-    switch (error.code) {
-      case "end_turn_archived_world":
-        return "This world is archived. End turn is unavailable.";
-      case "end_turn_running_transition":
-        return "Another end-turn transition is already running. Refresh the page before trying again.";
-      case "end_turn_stale_turn":
-        return "This turn has already changed. Refresh the page to review the latest world state.";
-      case "end_turn_transition_failed":
-        return "End turn could not be saved. Refresh the page before trying again.";
-      case "end_turn_unauthorized":
-        return "End turn is unavailable for this world.";
-    }
-  }
-
-  return "Try refreshing the page. If the problem continues, contact an administrator.";
 }
